All shadowban_threshold to be disabled
This allows the use of 0 to disable shadowban_threshold and allow PGScout to run without stopping due to errors.  I am seeing random errors as a matter of course and there is no mechanism to reset the error count.  Eventually all accounts get turned off for shadowbans but they are not actually shadowbanned